+++ src/main/java/dev/willyelton/crystal_tools/utils/TransferUtils.java
@@ -1,7 +1,10 @@
 package dev.willyelton.crystal_tools.utils;
 
+import net.minecraft.core.NonNullList;
+import net.minecraft.core.component.DataComponents;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.component.ItemContainerContents;
 import net.neoforged.neoforge.transfer.IndexModifier;
 import net.neoforged.neoforge.transfer.ResourceHandler;
 import net.neoforged.neoforge.transfer.item.ItemResource;
@@ -16,7 +19,8 @@ import java.util.List;
 public class TransferUtils {
     public static IndexModifier<ItemResource> playerIndexModifier(Inventory inventory) {
         return (index, resource, amount) -> {
-            ItemResource.of(inventory.getItem(index));
+            inventory.setItem(index, resource.toStack(amount));
+            inventory.setChanged();
         };
     }
 
@@ -34,7 +38,7 @@ public class TransferUtils {
         }
     }
 
-    // TODO: Doing this seems wrong
+    // TODO: Doing this seems wrong and doesn't work for some reason
     public static IndexModifier<ItemResource> indexModifier(ResourceHandler<ItemResource> handler) {
         return (index, resource, amount) -> {
             try (Transaction tx = Transaction.open(null)) {
@@ -43,9 +47,22 @@ public class TransferUtils {
                     handler.extract(index, invResource, handler.getCapacityAsInt(index, invResource), tx);
                 }
 
-                handler.insert(index, resource, amount, tx);
+                if (!resource.isEmpty()) {
+                    handler.insert(index, resource, amount, tx);
+                }
+
                 tx.commit();
             }
+        };
+    }
+
+    public static IndexModifier<ItemResource> itemstackIndexModifier(ItemStack stack) {
+        return (index, resource, amount) -> {
+            ItemContainerContents contents = stack.getOrDefault(DataComponents.CONTAINER, ItemContainerContents.EMPTY);
+            NonNullList<ItemStack> list = NonNullList.withSize(Math.max(contents.getSlots(), 256), ItemStack.EMPTY);
+            contents.copyInto(list);
+            list.set(index, resource.toStack(amount));
+            stack.set(DataComponents.CONTAINER, ItemContainerContents.fromItems(list));
         };
     }